move watchCache metrics to a seperate file
